+++ Widgets/FarmingBar_Bar.lua
@@ -89,13 +89,14 @@ local function anchor_OnEnter(self)
 
     local tooltip = widget:GetUserData("tooltip")
     if tooltip and not addon.DragFrame:GetObjective() then
-        addon.tooltip:SetScript("OnUpdate", function()
-            addon.tooltip:ClearLines()
-            addon.tooltip:SetOwner(self, "ANCHOR_BOTTOMRIGHT", 0, 0)
-            addon[tooltip](addon, widget, addon.tooltip)
-            addon.tooltip:Show()
+        local tooltipFrame = addon:GetDBValue("global", "settings.tooltips.useGameTooltip") and GameTooltip or addon.tooltip
+        tooltipFrame:SetScript("OnUpdate", function()
+            tooltipFrame:ClearLines()
+            tooltipFrame:SetOwner(self, "ANCHOR_BOTTOMRIGHT", 0, 0)
+            addon[tooltip](addon, widget, tooltipFrame)
+            tooltipFrame:Show()
         end)
-        addon.tooltip:GetScript("OnUpdate")()
+        tooltipFrame:GetScript("OnUpdate")()
     end
 end
 
@@ -107,9 +108,10 @@ local function anchor_OnLeave(self)
 
     local tooltip = widget:GetUserData("tooltip")
     if tooltip and not addon.DragFrame:GetObjective() then
-        addon.tooltip:ClearLines()
-        addon.tooltip:Hide()
-        addon.tooltip:SetScript("OnUpdate", nil)
+        local tooltipFrame = addon:GetDBValue("global", "settings.tooltips.useGameTooltip") and GameTooltip or addon.tooltip
+        tooltipFrame:ClearLines()
+        tooltipFrame:Hide()
+        tooltipFrame:SetScript("OnUpdate", nil)
     end
 end